New files

+++ src/kift_server.c
@@ -6,7 +6,7 @@
 /*   By: jkalia <jkalia@student.42.us.org>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/06/12 00:57:43 by jkalia            #+#    #+#             */
-/*   Updated: 2017/06/18 17:08:01 by gguiulfo         ###   ########.fr       */
+/*   Updated: 2017/06/26 22:08:02 by gguiulfo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,9 @@
 #include <fcntl.h>
 #include <stdlib.h>
 #include <unistd.h>
+#include <netdb.h>
+#include <sys/types.h>
+#include <errno.h>
 
 int	g_new_socket;
 
@@ -69,22 +72,34 @@ static void	recieve_wav(t_server *server, char *inbuffer)
 	kift_log(server->recognized, server->response);
 }
 
-static int	begin(t_server *server)
+static int	begin(int filedes)
 {
+	t_server server;
+	int		nbytes;
 	char	buffer[BUFFER];
+	static int flag = 0;
 
-	bzero(server, sizeof(t_server));
+	bzero(&server, sizeof(t_server));
 	bzero(buffer, BUFFER);
-	CHK1(read(g_new_socket, buffer, BUFFER) == -1, printf("ERROR READ"), -1);
+	nbytes = read(filedes, buffer, BUFFER);
+	if (nbytes < 0)
+	{
+		perror("read");
+		exit(EXIT_FAILURE);
+	}
 	printf("%s\n", buffer);
 	if (ISDIGIT(buffer[0]))
-		recieve_wav(server, buffer);
+		recieve_wav(&server, buffer);
 	else
 	{
 		printf("ERROR!");
 		exit(EXIT_FAILURE);
 	}
-	system("rm -f out.wav");
+	if (flag == 0)
+	{
+		system("rm -f out.wav");
+		flag = 1;
+	}
 	return (0);
 }
 
@@ -106,30 +121,142 @@ static int	check_port(int argc, char **argv)
 	return (tmp);
 }
 
-int			main(int argc, char **argv)
-{
-	struct sockaddr_storage	server_storage;
-	struct sockaddr_in		server_addr;
-	socklen_t				addr_size;
-	t_server				server;
-	int						welcome_socket;
+// int			main(int argc, char **argv)
+// {
+// 	struct sockaddr_storage	server_storage;
+// 	struct sockaddr_in		server_addr;
+// 	socklen_t				addr_size;
+// 	t_server				server;
+// 	int						welcome_socket;
+//
+// 	bzero(&server, sizeof(server));
+// 	welcome_socket = socket(PF_INET, SOCK_STREAM, 0);
+// 	server_addr.sin_family = AF_INET;
+// 	server_addr.sin_port = htons(check_port(argc, argv));
+// 	server_addr.sin_addr.s_addr = inet_addr("127.0.0.1");
+// 	memset(server_addr.sin_zero, '\0', sizeof(server_addr.sin_zero));
+// 	bind(welcome_socket, (struct sockaddr *)&server_addr, sizeof(server_addr));
+// 	listen(welcome_socket, 5) == 0 ? printf("Listening\n") : printf("Error\n");
+// 	addr_size = sizeof(server_storage);
+// 	while (1)
+// 	{
+// 		g_new_socket = accept(welcome_socket,
+// 					(struct sockaddr *)&server_storage, &addr_size);
+// 		printf("%s:%d connected\n", inet_ntoa(server_addr.sin_addr),
+// 					ntohs(server_addr.sin_port));
+// 		begin(&server);
+// 	}
+// 	return (0);
+// }
 
-	bzero(&server, sizeof(server));
-	welcome_socket = socket(PF_INET, SOCK_STREAM, 0);
-	server_addr.sin_family = AF_INET;
-	server_addr.sin_port = htons(check_port(argc, argv));
-	server_addr.sin_addr.s_addr = inet_addr("127.0.0.1");
-	memset(server_addr.sin_zero, '\0', sizeof(server_addr.sin_zero));
-	bind(welcome_socket, (struct sockaddr *)&server_addr, sizeof(server_addr));
-	listen(welcome_socket, 5) == 0 ? printf("Listening\n") : printf("Error\n");
-	addr_size = sizeof(server_storage);
+#define PORT    4242
+#define MAXMSG  512
+
+int	make_socket (uint16_t port)
+{
+	int sock;
+	struct sockaddr_in name;
+	/* Create the socket. */
+	sock = socket (PF_INET, SOCK_STREAM, 0);
+	if (sock < 0)
+	{
+		perror ("socket");
+		exit (EXIT_FAILURE);
+	}
+	/* Give the socket a name. */
+	name.sin_family = AF_INET;
+	name.sin_port = htons(port);
+	name.sin_addr.s_addr = htonl(INADDR_ANY);
+	if (bind (sock, (struct sockaddr *) &name, sizeof (name)) < 0)
+	{
+		perror("bind");
+		exit(EXIT_FAILURE);
+	}
+	return (sock);
+}
+
+int
+read_from_client (int filedes)
+{
+	char buffer[MAXMSG];
+	int nbytes;
+
+	nbytes = read (filedes, buffer, MAXMSG);
+	if (nbytes < 0)
+	{
+		/* Read error. */
+		perror ("read");
+		exit (EXIT_FAILURE);
+	}
+	else if (nbytes == 0)
+	/* End-of-file. */
+	return -1;
+	else
+	{
+		/* Data read. */
+		fprintf (stderr, "Server: got message: `%s'\n", buffer);
+		return 0;
+	}
+}
+
+int	main(void)
+{
+	int sock;
+	fd_set active_fd_set, read_fd_set;
+	int i;
+	struct sockaddr_in clientname;
+	size_t size;
+	/* Create the socket and set it up to accept connections. */
+	sock = make_socket(PORT);
+	if (listen (sock, 1) < 0)
+	{
+		perror ("listen");
+		exit (EXIT_FAILURE);
+	}
+	/* Initialize the set of active sockets. */
+	FD_ZERO (&active_fd_set);
+	FD_SET (sock, &active_fd_set);
 	while (1)
 	{
-		g_new_socket = accept(welcome_socket,
-					(struct sockaddr *)&server_storage, &addr_size);
-		printf("%s:%d connected\n", inet_ntoa(server_addr.sin_addr),
-					ntohs(server_addr.sin_port));
-		begin(&server);
+		/* Block until input arrives on one or more active sockets. */
+		read_fd_set = active_fd_set;
+		if (select (FD_SETSIZE, &read_fd_set, NULL, NULL, NULL) < 0)
+		{
+			perror ("select");
+			exit (EXIT_FAILURE);
+		}
+		/* Service all the sockets with input pending. */
+		for (i = 0; i < FD_SETSIZE; ++i)
+		if (FD_ISSET (i, &read_fd_set))
+		{
+			if (i == sock)
+			{
+				/* Connection request on original socket. */
+				int new;
+				size = sizeof (clientname);
+				new = accept(sock, (struct sockaddr *) &clientname, &size);
+				if (new < 0)
+				{
+					perror("accept");
+					exit(EXIT_FAILURE);
+				}
+				fprintf(stderr,
+					"Server: connect from host %s, port %hd.\n",
+					inet_ntoa(clientname.sin_addr),
+					ntohs(clientname.sin_port));
+					FD_SET(new, &active_fd_set);
+			}
+			else
+			{
+				/* Data arriving on an already-connected socket. */
+				// if (read_from_client (i) < 0)
+				// {
+					// bzero(&server, sizeof(t_server));
+					begin(i);
+					close(i);
+					FD_CLR(i, &active_fd_set);
+				// }
+			}
+		}
 	}
-	return (0);
 }
